requirements.txt
# QA
flake8
flake8-docstrings
isort

# App
tornado
pyserial
netifaces
pyyaml
